+++ packages/gallery/src/components/item/itemView.js
@@ -593,7 +593,7 @@ class ItemView extends GalleryComponent {
   }
   
   getItemContainerStyles() {
-    const { offset, style, styleParams } = this.props;
+    const { offset, style, styleParams, settings = {} } = this.props;
 
     const containerStyleByStyleParams = getContainerStyle(styleParams);
     const itemDoesntHaveLink = !this.itemHasLink(); //when itemClick is 'link' but no link was added to this specific item
@@ -602,18 +602,22 @@ class ItemView extends GalleryComponent {
       overflowY: styleParams.isSlideshow ? 'visible' : 'hidden',
       position: 'absolute',
       bottom: 'auto',
-      top: offset.top,
-      left: styleParams.isRTL ? 'auto' : offset.left,
-      right: !styleParams.isRTL ? 'auto' : offset.left,
-      width: style.width + style.infoWidth,
-      height: style.height + style.infoHeight,
       margin: styleParams.oneRow ? styleParams.imageMargin + 'px' : 0,
       cursor: styleParams.itemClick === CLICK_ACTIONS.NOTHING ||
       (styleParams.itemClick === CLICK_ACTIONS.LINK && itemDoesntHaveLink)
         ? 'default'
         : 'pointer'
     };
-    return { ...itemStyles, ...containerStyleByStyleParams };
+
+    const layoutStyles = settings.avoidInlineStyles ? {} : {
+      top: offset.top,
+      left: styleParams.isRTL ? 'auto' : offset.left,
+      right: !styleParams.isRTL ? 'auto' : offset.left,
+      width: style.width + style.infoWidth,
+      height: style.height + style.infoHeight,
+    };
+
+    return { ...itemStyles, ...layoutStyles, ...containerStyleByStyleParams };
   }
 
   getItemWrapperStyles() {