+++ packages/gallery/src/versionLogger.js
@@ -1,1 +1,1 @@
-try {console.debug('Pro Gallery Version: 2.0.13'); } catch (e) {}
+try {console.debug('Pro Gallery Version: 2.0.16'); } catch (e) {}